+++ pkg/common/models/balance.go
@@ -1,7 +1,7 @@
 package models
 
 type Balance struct {
-	Id       int    `json:"id" gorm:"primaryKey"`
-	Amount   int    `json:"amount"`
-	Currency string `json:"currency"`
+	Id       int     `json:"id" gorm:"primaryKey"`
+	Amount   float64 `json:"amount"`
+	Currency string  `json:"currency"`
 }
